hunter2-lessons/jwt-key-flaws/code/requirements.txt
flask
pyjwt
flask-restful
orator
flask-orator
bcrypt
pycrypto
cryptography
